Renames pwv_kpno/download_suomi_data.py -> pwv_kpno/download_pwv_data.py
